The `rosvgtree` crate was merged into `usvg-parser`.
There is no point in having it as a separate crate.

+++ crates/resvg/src/filter/color_matrix.rs
@@ -54,8 +54,8 @@ pub fn apply(matrix: &ColorMatrix, src: ImageRefMut) {
         }
         ColorMatrix::HueRotate(angle) => {
             let angle = angle.to_radians();
-            let a1 = angle.cos() as f32;
-            let a2 = angle.sin() as f32;
+            let a1 = angle.cos();
+            let a2 = angle.sin();
             let m = [
                 0.213 + 0.787 * a1 - 0.213 * a2,
                 0.715 - 0.715 * a1 - 0.715 * a2,
